move into new util file
